+++ core-services/egov-hrms/src/main/java/org/egov/hrms/service/IndividualService.java
@@ -2,6 +2,7 @@ package org.egov.hrms.service;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
@@ -13,6 +14,7 @@ import java.util.stream.Collectors;
 import digit.models.coremodels.AuditDetails;
 import digit.models.coremodels.user.enums.UserType;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.egov.common.contract.request.RequestInfo;
 import org.egov.common.models.core.Role;
 import org.egov.common.models.individual.Address;
@@ -23,8 +25,6 @@ import org.egov.common.models.individual.Individual;
 import org.egov.common.models.individual.IndividualBulkResponse;
 import org.egov.common.models.individual.IndividualRequest;
 import org.egov.common.models.individual.IndividualResponse;
-import org.egov.common.models.individual.IndividualSearch;
-import org.egov.common.models.individual.IndividualSearchRequest;
 import org.egov.common.models.individual.Name;
 import org.egov.common.models.individual.UserDetails;
 import org.egov.hrms.config.PropertiesManager;
@@ -33,6 +33,8 @@ import org.egov.hrms.utils.HRMSConstants;
 import org.egov.hrms.web.contract.User;
 import org.egov.hrms.web.contract.UserRequest;
 import org.egov.hrms.web.contract.UserResponse;
+import org.egov.hrms.web.models.IndividualSearch;
+import org.egov.hrms.web.models.IndividualSearchRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import static org.egov.hrms.utils.HRMSConstants.SYSTEM_GENERATED;
@@ -70,13 +72,20 @@ public class IndividualService implements UserService {
 
     @Override
     public UserResponse updateUser(UserRequest userRequest) {
-        IndividualRequest request = mapToIndividualRequest(userRequest);
+        IndividualSearchRequest individualSearchRequest = mapToIndividualSearchRequest(userRequest);
+        IndividualBulkResponse individualSearchResponse =
+                getIndividualResponse(userRequest.getUser().getTenantId(), individualSearchRequest);
+        UserResponse userResponse = null;
+        if (individualSearchResponse == null || individualSearchResponse.getIndividual() == null || individualSearchResponse.getIndividual().size() == 0) {
+            return userResponse;
+        }
+        Individual individual = individualSearchResponse.getIndividual().get(0);
+        IndividualRequest updateRequest = mapToIndividualUpdateRequest(individual, userRequest);
         StringBuilder uri = new StringBuilder();
         uri.append(propertiesManager.getIndividualHost());
         uri.append(propertiesManager.getIndividualUpdateEndpoint());
         IndividualResponse response = restCallRepository
-                .fetchResult(uri, request, IndividualResponse.class);
-        UserResponse userResponse = null;
+                .fetchResult(uri, updateRequest, IndividualResponse.class);
         if (response != null && response.getIndividual() != null) {
             log.info("response received from individual service");
             userResponse = mapToUserResponse(response);
@@ -84,18 +93,94 @@ public class IndividualService implements UserService {
         return userResponse;
     }
 
+    private IndividualRequest mapToIndividualUpdateRequest(Individual individual, UserRequest userRequest) {
+        Individual updatedIndividual = Individual.builder()
+                .id(individual.getId())
+                .userId(individual.getUserId())
+                .userUuid(individual.getUserUuid())
+                .isSystemUser(true)
+                .isSystemUserActive(userRequest.getUser().getActive())
+                .name(Name.builder()
+                        .givenName(userRequest.getUser().getName())
+                        .build())
+                .gender(Gender.fromValue(userRequest.getUser().getGender()))
+                .email(userRequest.getUser().getEmailId())
+                .mobileNumber(userRequest.getUser().getMobileNumber())
+                .dateOfBirth(convertMillisecondsToDate(userRequest.getUser().getDob()))
+                .tenantId(userRequest.getUser().getTenantId())
+                .address(Collections.singletonList(Address.builder()
+                        .type(AddressType.CORRESPONDENCE)
+                        .addressLine1(userRequest.getUser().getCorrespondenceAddress())
+                        .clientReferenceId(String.valueOf(UUID.randomUUID()))
+                        .isDeleted(Boolean.FALSE)
+                        .build()))
+                /*
+                 * FIXME (HCM specific change) clientReferenceId is the primary key in the individual table of the FrontEnd Worker Application's local database.
+                 */
+                // Generating a unique client reference ID using UUID
+                .clientReferenceId(individual.getClientReferenceId())
+                .userDetails(UserDetails.builder()
+                        .username(userRequest.getUser().getUserName())
+                        .password(userRequest.getUser().getPassword())
+                        .tenantId(userRequest.getUser().getTenantId())
+                        .roles(userRequest.getUser().getRoles().stream().map(role -> Role.builder()
+                                .code(role.getCode())
+                                .name(role.getName())
+                                .tenantId(userRequest.getUser().getTenantId())
+                                .description(role.getDescription())
+                                .build()).collect(Collectors.toList()))
+                        .userType(UserType.fromValue(userRequest.getUser().getType()))
+                        .build())
+                .isDeleted(Boolean.FALSE)
+                .clientAuditDetails(AuditDetails.builder()
+                        .createdBy(individual.getAuditDetails().getCreatedBy())
+                        .createdTime(individual.getAuditDetails().getCreatedTime())
+                        .lastModifiedBy(userRequest.getRequestInfo().getUserInfo().getUuid()).build())
+                .rowVersion(userRequest.getUser().getRowVersion())
+                .build();
+        return IndividualRequest.builder()
+                .requestInfo(userRequest.getRequestInfo())
+                .individual(updatedIndividual)
+                .build();
+    }
+
+    private IndividualSearchRequest mapToIndividualSearchRequest(UserRequest userRequest) {
+        return IndividualSearchRequest.builder()
+                .requestInfo(userRequest.getRequestInfo())
+                .individual(
+                        IndividualSearch.builder()
+                        .id(Collections.singletonList(userRequest.getUser().getUuid()))
+                        .userUuid(userRequest.getUser().getUserServiceUuid() != null ? Collections.singletonList(userRequest.getUser().getUserServiceUuid()) : null)
+                        .build()
+                )
+                .build();
+    }
+
     @Override
     public UserResponse getUser(RequestInfo requestInfo, Map<String, Object> userSearchCriteria ) {
+        String mobileNumber = (String) userSearchCriteria.get("mobileNumber");
+        String username = (String) userSearchCriteria.get(HRMSConstants.HRMS_USER_SEARCH_CRITERA_USERNAME);
+        List<String> mobileNumberList = null;
+        List<String> usernameList = null;
+        if(!StringUtils.isEmpty(mobileNumber)) {
+            mobileNumberList = Collections.singletonList(mobileNumber);
+        }
+        if(!StringUtils.isEmpty(username)) {
+            usernameList = Collections.singletonList(username);
+        }
         IndividualSearchRequest request = IndividualSearchRequest.builder()
                 .requestInfo(requestInfo)
                 .individual(IndividualSearch.builder()
-                        .mobileNumber((String) userSearchCriteria.get("mobileNumber"))
+                        .mobileNumber(
+                                mobileNumberList
+                        )
                         .id((List<String>) userSearchCriteria.get("uuid"))
                         .roleCodes((List<String>) userSearchCriteria.get("roleCodes"))
-                        .username((String) userSearchCriteria.get(HRMSConstants.HRMS_USER_SEARCH_CRITERA_USERNAME))
+                        .username(usernameList)
                         // given name
                         .individualName((String) userSearchCriteria
                                 .get(HRMSConstants.HRMS_USER_SEARCH_CRITERA_NAME))
+                        .type((String) userSearchCriteria.get(HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE_CODE))
                 .build())
                 .build();
         IndividualBulkResponse response = getIndividualResponse((String) userSearchCriteria