+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/individual/IndividualSearch.java
@@ -1,42 +1,38 @@
 package org.egov.common.models.individual;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.validation.annotation.Validated;
-
-import javax.validation.Valid;
-import javax.validation.constraints.Size;
 import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.annotations.ApiModel;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.DecimalMax;
+import jakarta.validation.constraints.DecimalMin;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.egov.common.models.core.EgovOfflineSearchModel;
+import org.egov.common.models.core.Exclude;
+import org.springframework.validation.annotation.Validated;
+
 /**
 * A representation of an Individual.
 */
-    @ApiModel(description = "A representation of an Individual.")
+@ApiModel(description = "A representation of an Individual.")
 @Validated
-@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2022-12-27T11:47:19.561+05:30")
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-    @JsonIgnoreProperties(ignoreUnknown = true)
-public class IndividualSearch   {
-    @JsonProperty("id")
-    private List<String> id = null;
-
+@SuperBuilder
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class IndividualSearch extends EgovOfflineSearchModel {
     @JsonProperty("individualId")
-    private String individualId = null;
-
-    @JsonProperty("clientReferenceId")
-    private List<String> clientReferenceId = null;
+    private List<String> individualId = null;
 
     @JsonProperty("name")
     @Valid
@@ -51,7 +47,7 @@ public class IndividualSearch   {
     private Gender gender = null;
 
     @JsonProperty("mobileNumber")
-    private String mobileNumber = null;
+    private List<String> mobileNumber = null;
 
     @JsonProperty("socialCategory")
     private String socialCategory = null;
@@ -79,13 +75,37 @@ public class IndividualSearch   {
     private List<String> roleCodes = null;
 
     @JsonProperty("username")
-    private String username;
+    private List<String> username;
 
     @JsonProperty("userId")
-    private Long userId;
+    private List<Long> userId;
 
     @JsonProperty("userUuid")
     @Size(min = 1)
     private List<String> userUuid;
+
+    @Exclude
+    @JsonProperty("latitude")
+    @DecimalMin("-90")
+    @DecimalMax("90")
+    private Double latitude;
+
+    @Exclude
+    @JsonProperty("longitude")
+    @DecimalMin("-180")
+    @DecimalMax("180")
+    private Double longitude;
+
+    /*
+     * @value unit of measurement in Kilometer
+     * */
+    @Exclude
+    @JsonProperty("searchRadius")
+    @DecimalMin("0")
+    private Double searchRadius;
+
+
+    @JsonProperty("type")
+    private String type;
 }