+++ health-services/stock/src/main/java/org/egov/stock/service/StockReconciliationService.java
@@ -10,6 +10,7 @@ import org.egov.common.validator.Validator;
 import org.egov.stock.config.StockReconciliationConfiguration;
 import org.egov.stock.repository.StockReconciliationRepository;
 import org.egov.stock.service.enrichment.StockReconciliationEnrichmentService;
+import org.egov.stock.validator.stockreconciliation.SrExistentEntityValidator;
 import org.egov.stock.validator.stockreconciliation.SrFacilityIdValidator;
 import org.egov.stock.validator.stockreconciliation.SrIsDeletedValidator;
 import org.egov.stock.validator.stockreconciliation.SrNonExistentValidator;
@@ -18,7 +19,7 @@ import org.egov.stock.validator.stockreconciliation.SrProductVariantIdValidator;
 import org.egov.stock.validator.stockreconciliation.SrReferenceIdValidator;
 import org.egov.stock.validator.stockreconciliation.SrRowVersionValidator;
 import org.egov.stock.validator.stockreconciliation.SrUniqueEntityValidator;
-import org.egov.stock.web.models.StockReconciliationSearchRequest;
+import org.egov.common.models.stock.StockReconciliationSearchRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ReflectionUtils;
 
@@ -56,6 +57,7 @@ public class StockReconciliationService {
 
     private final Predicate<Validator<StockReconciliationBulkRequest, StockReconciliation>> isApplicableForCreate =
             validator -> validator.getClass().equals(SrProductVariantIdValidator.class)
+                    || validator.getClass().equals(SrExistentEntityValidator.class)
                     || validator.getClass().equals(SrFacilityIdValidator.class)
                     || validator.getClass().equals(SrReferenceIdValidator.class);