+++ health-services/transformer/src/main/java/org/egov/transformer/service/ProjectStaffTransformationService.java
@@ -5,7 +5,7 @@ import org.egov.common.models.project.ProjectStaff;
 import org.egov.transformer.config.TransformerProperties;
 import org.egov.transformer.enums.Operation;
 import org.egov.transformer.models.downstream.ProjectStaffIndexV1;
-import org.egov.transformer.producer.Producer;
+import org.egov.common.producer.Producer;
 import org.egov.transformer.service.transformer.Transformer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -67,7 +67,7 @@ public abstract class ProjectStaffTransformationService implements Transformatio
         @Override
         public List<ProjectStaffIndexV1> transform(ProjectStaff projectStaff) {
             Map<String, String> boundaryLabelToNameMap = projectService
-                    .getBoundaryLabelToNameMapByProjectId(projectStaff.getProjectId(), projectStaff.getTenantId());
+                    .getBoundaryCodeToNameMapByProjectId(projectStaff.getProjectId(), projectStaff.getTenantId());
             log.info("boundary labels {}", boundaryLabelToNameMap.toString());
             return Collections.singletonList(ProjectStaffIndexV1.builder()
                     .id(projectStaff.getId())