+++ health-services/stock/CHANGELOG.md
@@ -1,5 +1,11 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.3 - 2024-05-29
+- Upgraded to Core 2.9LTS
+- Upgraded to health models 1.0.20 and health common 1.0.16
+- Boundary v2 Integration
+- MDMS v2 integration
+
 ## 1.1.2 - 2024-02-26
 - Enhance inventory flow with sender id and receiver id added.